installing-ndctl: fix formatting in the troubleshooting section
The snapshot of the ndctl README that this was created from had a few
formatting errors, one of which, the depmod configuration would cause a
user to hit a failure if copy-pasting (missing newline after the
device-dax override).

Fix up the troubleshooting section to use the same preformatted text
directive as the rest of the document.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>